requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
click==8.0.1
elasticsearch==7.13.2
elasticsearch-dsl==7.3.0
fastapi==0.65.2
geojson-pydantic==0.3.0
h11==0.12.0
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.18.0
six==1.16.0
-e git+https://github.com/rsmith013/stac-fastapi@885faee572c1884337e5dd535ff4d68610a4625f#egg=stac_fastapi.api&subdirectory=stac_fastapi/api
-e git+https://github.com/cedadev/stac-fastapi-elasticsearch.git@f198b4d8f75ed7febe86bb86fc2db9aaf65877b9#egg=stac_fastapi.elasticsearch
-e git+https://github.com/rsmith013/stac-fastapi@885faee572c1884337e5dd535ff4d68610a4625f#egg=stac_fastapi.extensions&subdirectory=stac_fastapi/extensions
-e git+https://github.com/rsmith013/stac-fastapi@885faee572c1884337e5dd535ff4d68610a4625f#egg=stac_fastapi.types&subdirectory=stac_fastapi/types
-e git+https://github.com/stac-utils/stac-pydantic.git@200d39dbe927a5664bd9feda7cff8dfa356de605#egg=stac_pydantic
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
